+++ front/src/__tests__/atoms/BackButton.test.tsx
@@ -1,9 +1,7 @@
-import { vi } from "vitest"
-import { fireEvent, render, screen } from "@testing-library/react"
-import { BackButton } from "../../components/atoms/BackButton"
-import { waitFor } from '../test-utils'
+import { vi } from 'vitest'
+import { fireEvent, render, screen, waitFor } from '../test-utils'
+import { BackButton } from '../../components/atoms/BackButton'
 
-const mockClick = vi.fn()
 const navigate = vi.fn()
 
 vi.mock('react-router-dom', async () => {
@@ -16,23 +14,23 @@ vi.mock('react-router-dom', async () => {
   }
 })
 
-describe('BackButton', () => {
+describe('BackButton component', () => {
   it('renders correctly', () => {
-    render(
-      <BackButton onClick={ mockClick } />
-    )
-    const button = screen.getByText(/volver/i)
+    render(<BackButton />)
+    const button = screen.getByRole('button', {
+      name: /arrow_back_ios torna/i,
+    })
     expect(button).toBeInTheDocument()
   })
 
   it('triggers navigation to the previous page on click event', async () => {
-    render(
-      <BackButton onClick={ mockClick } />
-    )
-    const button = screen.getByText(/volver/i)
+    render(<BackButton />)
+    const button = screen.getByRole('button', {
+      name: /arrow_back_ios torna/i,
+    })
     fireEvent.click(button)
     await waitFor(() => {
-    expect(navigate).toHaveBeenCalledWith(-1)
+      expect(navigate).toHaveBeenCalledWith(-1)
     })
   })
 })