+++ front/src/__tests__/atoms/Counter.test.tsx
@@ -3,18 +3,48 @@ import { Counter } from '../../components/atoms'
 
 describe('Counter', () => {
   it('renders correctly with given props', () => {
+    render(<Counter number={37} text="Comments" icon="test" isError={false} />)
+
+    const number = screen.getByText(37)
+    expect(number).toBeInTheDocument()
+    const text = screen.getByText(/comments/i)
+    expect(text).toBeInTheDocument()
+    const icon = screen.getByText(/test/i)
+    expect(icon).toBeInTheDocument()
+    expect(screen.queryByRole('status')).not.toBeInTheDocument()
+  })
+
+  it('renders correctly when number is 0', () => {
+    render(<Counter number={0} text="Comments" icon="test" isError={false} />)
+
+    const number = screen.getByText(0)
+    expect(number).toBeInTheDocument()
+    const text = screen.getByText(/comments/i)
+    expect(text).toBeInTheDocument()
+    const icon = screen.getByText(/test/i)
+    expect(icon).toBeInTheDocument()
+    expect(screen.queryByRole('status')).not.toBeInTheDocument()
+  })
+
+  it('renders spinner if there is no number', () => {
     render(
-      <Counter
-        number={ 37 }
-        text='Comments'
-        icon='test'    
-      />
+      <Counter number={undefined} text="Comments" icon="test" isError={false} />
     )
-      const number: HTMLParagraphElement = screen.getByText(37)
-      expect(number).toBeInTheDocument()
-      const text: HTMLParagraphElement = screen.getByText(/comments/i)
-      expect(text).toBeInTheDocument()
-      const icon: HTMLSpanElement = screen.getByText(/test/i)
-      expect(icon).toBeInTheDocument()
+
+    expect(screen.getByRole('status')).toBeInTheDocument()
+    const text = screen.getByText(/comments/i)
+    expect(text).toBeInTheDocument()
+    const icon = screen.getByText(/test/i)
+    expect(icon).toBeInTheDocument()
+  })
+
+  it('renders not available text if there is an error with data', () => {
+    render(<Counter number={undefined} text="Comments" icon="test" isError />)
+
+    expect(screen.getByText('n/d')).toBeInTheDocument()
+    const text = screen.getByText(/comments/i)
+    expect(text).toBeInTheDocument()
+    const icon = screen.getByText(/test/i)
+    expect(icon).toBeInTheDocument()
   })
 })