+++ front/src/components/atoms/BackButton.tsx
@@ -1,20 +1,37 @@
 import { FC, HTMLAttributes } from 'react'
 import { useNavigate } from 'react-router-dom'
+import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
 import { Icon } from '.'
-import { font, dimensions } from '../../styles'
+import { font, device, dimensions } from '../../styles'
 
 const ButtonStyled = styled.button`
-  margin: ${dimensions.spacing.none};
+  margin-left: 0;
   font-weight: ${font.medium};
   display: flex;
   align-items: center;
   position: absolute;
   top: ${dimensions.spacing.base};
-  left: ${dimensions.spacing.base};
+  left: ${dimensions.spacing.xxxs};
   border: none;
   background-color: transparent;
-  cursor: pointer;
+
+  @media only ${device.Tablet} {
+    margin-left: 2%;
+    cursor: pointer;
+    position: static;
+    top: 0;
+    left: 0;
+  }
+
+  @media only ${device.Laptop} {
+    margin-left: 5%;
+  }
+
+  @media only ${device.Desktop} {
+    margin-left: 7%;
+  }
+
   &:hover {
     opacity: 0.7;
   }
@@ -24,6 +41,7 @@ type TButton = HTMLAttributes<HTMLButtonElement>
 
 export const BackButton: FC<TButton> = () => {
   const navigate = useNavigate()
+  const { t } = useTranslation()
   const handlePrevPage = () => {
     navigate(-1)
   }
@@ -34,7 +52,7 @@ export const BackButton: FC<TButton> = () => {
         wght={700}
         style={{ fontSize: `${font.base}` }}
       />
-      Volver
+      {t('Volver')}
     </ButtonStyled>
   )
 }