+++ front/src/components/organisms/Navbar.tsx
@@ -37,7 +37,7 @@ const IconStyled = styled.div`
   padding: 6px;
   width: 3rem;
   height: ${dimensions.spacing.xxl};
-  border-radius: 20%;
+  border-radius: ${dimensions.borderRadius.base};
   background-color: ${colors.white};
   justify-content: center;
   align-items: center;
@@ -93,7 +93,11 @@ export const Navbar = ({ toggleModal, handleAccessModal }: TNavbar) => {
   }
 
   const location = useLocation()
-  const shouldRenderIcons = useMemo(() => location.pathname !== '/', [location])
+
+  const shouldRenderIcons = useMemo(() => {
+    const excludedPaths = ['/', '/profile']
+    return !excludedPaths.includes(location.pathname)
+  }, [location])
 
   return (
     <>