+++ sso/src/schemas/auth/loginSchema.ts
@@ -1,3 +1,3 @@
-import { userSchema } from '../userSchema'
+import { userSchema } from './userSchema'
 
 export const loginSchema = userSchema.pick({ dni: true, password: true })