+++ back/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0] - 2023-11-21
+
+### Changed
+
+- Added search query param to get resources endpoint.
+
+## [0.2.0] - 2023-11-20
+
+### Changed
+
+- Modify resources get controller to accept an array of topicIds.
+
 ## [0.1.0] - 2023-11-14
 
 ### Added