+++ front/CHANGELOG.md
@@ -2,6 +2,49 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.1] - 2023-11-21
+
+### Fixed
+
+- Updated border-radius to use fixed values from dimensions.
+
+## [0.4.0] - 2023-11-21
+
+### Changed
+
+- Updated api url to itawiki
+
+### Added
+
+- Laravel logo to Categories list
+- Category name to Category page (translation and test)
+
+## [0.3.0] - 2023-11-20
+
+### Added
+
+- Resources and favorites widgets to User Profile
+
+### Changed
+
+- User Profile UI
+- SortOrder type to avoid repetitions
+
+### Fixed
+
+- My Resources widget in Category page
+
+## [0.2.1] - 2023-11-16
+
+### Changed
+
+- Refactored votes/dates logic into a single component (VotesDateController).
+
+### Fixed
+
+- Fixed 'Votos' & 'Fecha' visual inconsistencies (arrow icon asc/desc, initial state).
+- Incorporated missing translations.
+
 ## [0.1.1] - 2023-11-15
 
 ### Fixed